move telemetry out of cache and into inmemorycache
